Refactor UI into feature modules
